Promote ov006/dScMgJump2_c to a single translation unit
Consolidates the 29 legacy per-function sources for the ov006 minigame
scene class dScMgJump2_c into one file, src/actors/d_sc_mg_jump2.cpp,
covering the whole .text run 0x020eebe8..0x020efc08.

The destructor is defined in the class body. D1 sits at 0x020eebe8, BELOW
D0 at 0x020eec9c; an out-of-line member definition makes mwccarm emit the
group as D0-then-D1 plus a homeless D2, and a single TU spanning the run
cannot then be linked. In-class emits D1-then-D0 and no D2. The key
function moves with it to OnYoshiTryEat (ov006:0x020efaa8), defined only
here, so _ZTV12dScMgJump2_c is still emitted by exactly one object.

The vptr store spells the address point as &_ZTV12dScMgJump2_c[2]:
objisolate refused the bare symbol with "unexpected reloc type=2 addend=0"
(tools/objisolate.py:472). Three DIFFs in the water-column helpers were a
struct-copy scheduling mismatch -- a two-word struct with two named int
fields compiles to interleaved ldr/str, while the ROM emits ldr/ldr/str/str,
which is what an array-member struct produces.

The manifest gains a 13-entry compiler_only_output block licensing the
class's own _ZTI/_ZTS/_ZTV plus the ten inherited base-RTTI deadstrip-data
records. 0x0213ccbc holds the string 12dScMgJump2_c, so the class name is
transcribed from the cartridge, not coined. Eight of the twelve tubuild
CONFLICT notes are marked RESOLVED and their markers removed: the TU now
carries one declaration per symbol, typed to what the ROM bytes require.

Gates: tubuild verify 29/29 MATCH, objisolate clean, all 29 functions in
ROM-ascending order; linkcheck SCRATCH-LINK-VERIFIED with the ROM
byte-identical to the stock control; rombuild 106/106 exact, 100.000000%,
mismatching 0, run again after promotion; romdata_check 7 VERIFIED / 6
PARTIAL / 0 DIFFERS; check_tubuild_conflicts and check_dead_references
clean. dsd check symbols --fail still fails on the same 9 pre-existing
baseline errors -- 0 new, and it is not green.
