+++ attribution.json
@@ -1412,6 +1412,35 @@
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c6RenderEv": "github-actions[bot]",
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c8BehaviorEv": "github-actions[bot]",
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c13InitResourcesEv": "github-actions[bot]",
-    "src/actors/d_a_obj_emm_log.cpp#daObjEmmLog_c_Spawn": "tangosdev"
+    "src/actors/d_a_obj_emm_log.cpp#daObjEmmLog_c_Spawn": "tangosdev",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_cD1Ev": "tangosdev",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_cD0Ev": "tangosdev",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eed64": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eed68": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eedc8": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eee3c": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eeef4": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eef40": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eef58": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eef90": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020eeff0": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef05c": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef0d4": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c16CleanupResourcesEv": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c6RenderEv": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef2b8": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef47c": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef480": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef4ec": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef580": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef5ac": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef768": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef794": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#func_ov006_020ef7f8": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c13OnTurnIntoEggEi": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c13OnYoshiTryEatEi": "github-actions[bot]",
+    "src/actors/d_sc_mg_jump2.cpp#dScMgJump2_c_Spawn": "github-actions[bot]"
   }
 }
