+++ config/arm9/overlays/ov006/delinks.txt
@@ -3206,121 +3206,9 @@ src/MgBounceAndPounce_Spawn.cpp:
     complete
     .text start:0x020eeafc end:0x020eebe8
 
-src/_ZN12dScMgJump2_cD1Ev.cpp:
+src/actors/d_sc_mg_jump2.cpp:
     complete
-    .text start:0x020eebe8 end:0x020eec9c
-
-src/_ZN12dScMgJump2_cD0Ev.cpp:
-    complete
-    .text start:0x020eec9c end:0x020eed64
-
-src/func_ov006_020eed64.c:
-    complete
-    .text start:0x020eed64 end:0x020eed68
-
-src/func_ov006_020eed68.c:
-    complete
-    .text start:0x020eed68 end:0x020eedc8
-
-src/func_ov006_020eedc8.c:
-    complete
-    .text start:0x020eedc8 end:0x020eee3c
-
-src/func_ov006_020eee3c.c:
-    complete
-    .text start:0x020eee3c end:0x020eeef4
-
-src/func_ov006_020eeef4.c:
-    complete
-    .text start:0x020eeef4 end:0x020eef40
-
-src/func_ov006_020eef40.c:
-    complete
-    .text start:0x020eef40 end:0x020eef58
-
-src/func_ov006_020eef58.c:
-    complete
-    .text start:0x020eef58 end:0x020eef90
-
-src/func_ov006_020eef90.c:
-    complete
-    .text start:0x020eef90 end:0x020eeff0
-
-src/func_ov006_020eeff0.c:
-    complete
-    .text start:0x020eeff0 end:0x020ef05c
-
-src/func_ov006_020ef05c.c:
-    complete
-    .text start:0x020ef05c end:0x020ef0d4
-
-src/func_ov006_020ef0d4.c:
-    complete
-    .text start:0x020ef0d4 end:0x020ef110
-
-src/_ZN12dScMgJump2_c16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x020ef110 end:0x020ef148
-
-src/_ZN12dScMgJump2_c6RenderEv.cpp:
-    complete
-    .text start:0x020ef148 end:0x020ef2b8
-
-src/func_ov006_020ef2b8.c:
-    complete
-    .text start:0x020ef2b8 end:0x020ef3e0
-
-src/_ZN12dScMgJump2_c8BehaviorEv.cpp:
-    complete
-    .text start:0x020ef3e0 end:0x020ef47c
-
-src/func_ov006_020ef47c.c:
-    complete
-    .text start:0x020ef47c end:0x020ef480
-
-src/func_ov006_020ef480.c:
-    complete
-    .text start:0x020ef480 end:0x020ef4ec
-
-src/func_ov006_020ef4ec.c:
-    complete
-    .text start:0x020ef4ec end:0x020ef580
-
-src/func_ov006_020ef580.c:
-    complete
-    .text start:0x020ef580 end:0x020ef5ac
-
-src/func_ov006_020ef5ac.c:
-    complete
-    .text start:0x020ef5ac end:0x020ef768
-
-src/func_ov006_020ef768.c:
-    complete
-    .text start:0x020ef768 end:0x020ef794
-
-src/func_ov006_020ef794.c:
-    complete
-    .text start:0x020ef794 end:0x020ef7f8
-
-src/func_ov006_020ef7f8.c:
-    complete
-    .text start:0x020ef7f8 end:0x020ef834
-
-src/_ZN12dScMgJump2_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x020ef834 end:0x020efa84
-
-src/_ZN12dScMgJump2_c13OnTurnIntoEggEi.cpp:
-    complete
-    .text start:0x020efa84 end:0x020efaa8
-
-src/_ZN12dScMgJump2_c13OnYoshiTryEatEi.cpp:
-    complete
-    .text start:0x020efaa8 end:0x020efaf0
-
-src/dScMgJump2_c_Spawn.c:
-    complete
-    .text start:0x020efaf0 end:0x020efc08
+    .text start:0x020eebe8 end:0x020efc08
 
 src/func_ov006_020efc08.c:
     complete
