+++ config/converted-backslide-exceptions.jsonl
@@ -175,3 +175,4 @@
 {"path": "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD0Ev", "reason": "ov029/daObjWc_Obj03_c promoted to src/actors/d_a_obj_wc_obj03.cpp: D1/D0 carried as // @symbol mangled bodies for the ROM emission order (D1 below D0), costing the no-raw-offset and no-mangled-refs criteria for those two functions"}
 {"path": "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD1Ev", "reason": "ov029/daObjWc_Obj03_c promoted to src/actors/d_a_obj_wc_obj03.cpp: D1/D0 carried as // @symbol mangled bodies for the ROM emission order (D1 below D0), costing the no-raw-offset and no-mangled-refs criteria for those two functions"}
 {"path": "src/__sinit_ov044_02111314.c", "reason": "ov044 compiler-generated __sinit is now emitted by its exact owning TU; the standalone intake file was not an original source unit"}
+{"path": "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c16CleanupResourcesEv", "reason": "ov006/dScMgJump2_c promoted to a single TU (src/actors/d_sc_mg_jump2.cpp); 29/29 MATCH, rombuild 106/106"}
