+++ config/converted-baseline.json
@@ -7,7 +7,7 @@
     "no_codegen_trick",
     "no_mangled_refs"
   ],
-  "count": 2589,
+  "count": 2590,
   "converted": [
     "src/ARMMathLoadState.c",
     "src/ARMMathSaveState.c",
@@ -588,9 +588,6 @@
     "src/_ZN12dScMg3DEsp_c16CleanupResourcesEv.cpp",
     "src/_ZN12dScMgAmida_c5Unk36Ev.cpp",
     "src/_ZN12dScMgAmida_c9Virtual8CEv.cpp",
-    "src/_ZN12dScMgJump2_c13OnTurnIntoEggEi.cpp",
-    "src/_ZN12dScMgJump2_c13OnYoshiTryEatEi.cpp",
-    "src/_ZN12dScMgJump2_c16CleanupResourcesEv.cpp",
     "src/_ZN12dScMgLuigi_c21AfterCleanupResourcesEj.cpp",
     "src/_ZN12dScMgLuigi_c6RenderEv.cpp",
     "src/_ZN12dScMgLuigi_c8BehaviorEv.cpp",
@@ -2471,6 +2468,10 @@
     "src/actors/d_a_warpkun.cpp#_ZN11daWarpkun_c8BehaviorEv",
     "src/actors/d_a_warpkun.cpp#_ZN11daWarpkun_cD0Ev",
     "src/actors/d_a_warpkun.cpp#_ZN11daWarpkun_cD1Ev",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c13OnTurnIntoEggEi",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_c13OnYoshiTryEatEi",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_cD0Ev",
+    "src/actors/d_sc_mg_jump2.cpp#_ZN12dScMgJump2_cD1Ev",
     "src/actors/daBar_c.cpp#_ZN7daBar_c16CleanupResourcesEv",
     "src/actors/daBar_c.cpp#_ZN7daBar_c16OnPendingDestroyEv",
     "src/actors/daBar_c.cpp#_ZN7daBar_c6RenderEv",
