+++ include/dScMgJump2_c.h
@@ -24,8 +24,9 @@
  * fix the model's offset and extent without pretending the other element
  * type names are known.
  *
- * THE DESTRUCTOR IS NOT DEFINED INLINE -- a leaf. No separate operator
- * delete is needed: dScMgD3DBase_c, the immediate base, provides one. */
+ * THE DESTRUCTOR IS DEFINED IN THE CLASS BODY, for the emission-order reason
+ * spelled out at the definition below. No separate operator delete is needed:
+ * dScMgD3DBase_c, the immediate base, provides one. */
 #ifndef DSCMGJUMP2_C_H
 #define DSCMGJUMP2_C_H
 #include "dScMgD3DBase_c.h"
@@ -37,7 +38,25 @@ extern "C" void func_ov006_020c6f3c(void);
 extern "C" void func_ov006_020eed64(void);
 
 struct dScMgJump2_c : dScMgD3DBase_c {
-    virtual ~dScMgJump2_c();
+    /* DEFINED IN THE CLASS BODY, DELIBERATELY -- not a style choice.
+       D1 sits at 0x020eebe8, BELOW D0 at 0x020eec9c. An out-of-line member
+       definition makes mwccarm 2004/b56 emit the destructor group as
+       D0-then-D1 plus a homeless D2, and a single translation unit covering
+       the whole .text run 0x020eebe8..0x020efc08 cannot then be linked: the
+       functions are not emitted in ROM address order. Defining it here emits
+       D1 before D0, emits no D2 at all, and costs nothing else -- it is still
+       the same body, and it is still not inlined into callers because a
+       virtual destructor is always reached through the vtable.
+       It does give up the key function: with this inline, the first virtual
+       DECLARED but not defined here is OnYoshiTryEat below, which lives in
+       src/actors/d_sc_mg_jump2.cpp, so _ZTV12dScMgJump2_c is still emitted
+       by that one translation unit and by no other. */
+    virtual ~dScMgJump2_c()
+    {
+        _ZN5ModelD1Ev((char *)this + 0x5a14);
+        __destroy_arr(mArray3, 0x10, 0x24, (void *)func_ov006_020eed64);
+        __destroy_arr(mArray2, 6, 0xf0, (void *)func_ov006_020c6f3c);
+    }
     virtual void OnYoshiTryEat(int arg);               /* slot 18 */
     virtual int  OnTurnIntoEgg(int mode);              /* slot 19 */
