--- src/_ZN12dScMgJump2_c13OnTurnIntoEggEi.cpp
@@ -1,14 +1,0 @@
-//cpp
-// @symbol _ZN12dScMgJump2_c13OnTurnIntoEggEi
-// recovered name: dScMgJump2_c_OnTurnIntoEgg
-/* recovered: renamed to Class_Method, declarations from a shared header */
-#include "decl_common.h"
-#include "dScMgJump2_c.h"
-/* recovered: renamed to Class_Method */
-/* dScMgJump2_c::OnTurnIntoEgg - recovered from vtable slot identity */
-
-int dScMgJump2_c::OnTurnIntoEgg(int /* mode */)
-{
-    func_ov006_020c8a9c(0, 0);
-    return 1;
-}
