--- src/_ZN12dScMgJump2_c13OnYoshiTryEatEi.cpp
@@ -1,20 +1,0 @@
-//cpp
-// @symbol _ZN12dScMgJump2_c13OnYoshiTryEatEi
-// recovered name: dScMgJump2_c_OnYoshiTryEat_020efaa8
-/* recovered: renamed to Class_Method, declarations from a shared header */
-#include "decl_common.h"
-#include "dScMgJump2_c.h"
-/* recovered: renamed to Class_Method, vtable slot 18 -- an override of
-   dScMgBase_c::OnYoshiTryEat(int). The signature must repeat the base
-   declaration exactly, or mwcc appends a slot instead of overriding. */
-void dScMgJump2_c::OnYoshiTryEat(int /* arg */)
-{
-    char *c = (char *)this;
-
-  func_ov006_020c72b4();
-  func_ov006_020c719c(0, 0);
-  data_ov006_02140328 = 3;
-  func_ov006_020c44b4(0, 0);
-  func_ov006_020eeff0();
-  func_ov006_020ef7f8(c);
-}
