+++ src/actors/d_sc_mg_jump2.cpp
@@ -1,0 +1,930 @@
+//cpp
+/* SHADOW translation unit -- generated by tools/tubuild.py create.
+ * ov006/dScMgJump2_c  (29 function(s))
+ *
+ * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
+ * build. It is a STARTING POINT (plan sec 7.3): local shadow declarations
+ * below were carried verbatim from the legacy files, not reconciled against
+ * real project headers -- that judgement call is left to a human/LLM review,
+ * the way pilot #1 reconciled PoleLift::Render and ::CleanupResources by hand
+ * (see notes/tu-reconstruction-pilot-report.md sec 5.2).
+ *
+ * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
+ * emits one .text section per function, in the REVERSE of source order, so
+ * the highest-address ROM function is written FIRST here. Do not reorder;
+ * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
+ * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
+ *
+ * Assembled from these legacy one-function sources (ROM address order):
+ *   [0] 0x020eebe8  src/_ZN12dScMgJump2_cD1Ev.cpp
+ *   [1] 0x020eec9c  src/_ZN12dScMgJump2_cD0Ev.cpp
+ *   [2] 0x020eed64  src/func_ov006_020eed64.c
+ *   [3] 0x020eed68  src/func_ov006_020eed68.c
+ *   [4] 0x020eedc8  src/func_ov006_020eedc8.c
+ *   [5] 0x020eee3c  src/func_ov006_020eee3c.c
+ *   [6] 0x020eeef4  src/func_ov006_020eeef4.c
+ *   [7] 0x020eef40  src/func_ov006_020eef40.c
+ *   [8] 0x020eef58  src/func_ov006_020eef58.c
+ *   [9] 0x020eef90  src/func_ov006_020eef90.c
+ *   [10] 0x020eeff0  src/func_ov006_020eeff0.c
+ *   [11] 0x020ef05c  src/func_ov006_020ef05c.c
+ *   [12] 0x020ef0d4  src/func_ov006_020ef0d4.c
+ *   [13] 0x020ef110  src/_ZN12dScMgJump2_c16CleanupResourcesEv.cpp
+ *   [14] 0x020ef148  src/_ZN12dScMgJump2_c6RenderEv.cpp
+ *   [15] 0x020ef2b8  src/func_ov006_020ef2b8.c
+ *   [16] 0x020ef3e0  src/_ZN12dScMgJump2_c8BehaviorEv.cpp
+ *   [17] 0x020ef47c  src/func_ov006_020ef47c.c
+ *   [18] 0x020ef480  src/func_ov006_020ef480.c
+ *   [19] 0x020ef4ec  src/func_ov006_020ef4ec.c
+ *   [20] 0x020ef580  src/func_ov006_020ef580.c
+ *   [21] 0x020ef5ac  src/func_ov006_020ef5ac.c
+ *   [22] 0x020ef768  src/func_ov006_020ef768.c
+ *   [23] 0x020ef794  src/func_ov006_020ef794.c
+ *   [24] 0x020ef7f8  src/func_ov006_020ef7f8.c
+ *   [25] 0x020ef834  src/_ZN12dScMgJump2_c13InitResourcesEv.cpp
+ *   [26] 0x020efa84  src/_ZN12dScMgJump2_c13OnTurnIntoEggEi.cpp
+ *   [27] 0x020efaa8  src/_ZN12dScMgJump2_c13OnYoshiTryEatEi.cpp
+ *   [28] 0x020efaf0  src/dScMgJump2_c_Spawn.c
+ */
+
+/* TUBUILD NOTE -- #pragma directive(s) were present in the legacy sources
+ * of this TU. `#pragma long_calls` is POSITIONAL in mwccarm 2004/b56 and is
+ * carried verbatim before its own member below, bracketed with `off` so it
+ * cannot leak into later members (dropping it silently costs the pooled
+ * cross-overlay tail-call -- a byte diff; see ShutterBob in ov014).
+ * Any OTHER pragma is FILE-GLOBAL last-wins (opt_propagation,
+ * optimize_for_size) and is still left out: carried into a merged TU it
+ * would silently recompile every other member. Decide those by hand:
+ *   func_ov006_020ef5ac: #pragma opt_propagation off   [NOT carried -- review]
+ */
+
+/* Includes: union of the legacy files', first-seen in ROM-ascending
+ * processing order. NOT verified for header ordering constraints (e.g. a
+ * common.h-before-X rule) -- watch for new compile errors after this. */
+#include "dScMgJump2_c.h"
+#include "SharedFilePtr.h"
+#include "types.h"
+#include "decl_common.h"
+
+/* Local shadow declarations carried from the legacy files verbatim.
+ * NOT reconciled against real project headers -- check include/*.h for
+ * each of these before compiling; a real header should usually win. */
+/* shadow struct 'Node' */
+struct Node { struct Node* next; };
+
+/* shadow struct 'Elem' */
+struct Elem { char _0[0x20]; short f20; short _22; };
+
+/* shadow struct 'M48' */
+struct M48 { int w[12]; };
+
+/* shadow struct 'S3' */
+struct S3 { int w[3]; };
+
+/* shadow struct 'Vector3_16f' */
+struct Vector3_16f;
+
+/* shadow typedef 'void' */
+typedef void (dScMgJump2_c::*dScMgJump2_cState)();
+
+/* shadow struct 'G2' */
+struct G2 { int w[2]; };
+
+/* shadow typedef 'u8' */
+typedef unsigned char u8;
+
+/* shadow typedef 'u16' */
+typedef unsigned short u16;
+
+/* shadow typedef 's16' */
+typedef short s16;
+
+/* shadow struct 'Base' */
+struct Base {
+    virtual void m00();
+    virtual void m04();
+    virtual void m08();
+    virtual void m0c();
+    virtual void m10();
+    virtual void m14();
+    virtual void m18();
+    virtual void m1c();
+    virtual void m20();
+    virtual void m24();
+    virtual void m28();
+    virtual void m2c();
+    virtual void m30();
+    virtual void m34();
+    virtual void m38();
+    virtual void m3c();
+    virtual void m40();
+    virtual void m44();
+    virtual void m48(int x);
+};
+
+extern "C" {
+extern int _Z15ApproachLinear2Rsss(short* a, short b, short c);
+extern void _Z14ApproachLinearRiii(int* a, int b, int c);
+extern void func_ov006_020eeef4(struct Node* node);
+extern void func_0203d630(int* p, int m);
+extern struct Node* data_ov006_021421c0;
+extern void func_ov006_020eed68(char *c);
+extern void func_ov006_020eedc8(char *p);
+extern int data_ov006_021421bc;
+extern struct Elem *data_ov006_021421b0;
+extern void func_ov006_020eee3c(int *thiz, int a1, int a2, int a3);
+extern void *data_ov006_021421b8;
+void func_ov004_020ad90c(void);
+void Matrix4x3_FromTranslation(void *m, int x, int y, int z);
+int GetGameLanguage(void);
+void DrawOamSprite(void *a0, void *a1, int a2, void *a3);
+void RenderOamMainScreen(int a0, int a1, int a2, int a3, int a4);
+extern int data_020a0e68;
+unsigned int _ZN8Particle6System17NewUnkCallback818Ejj5Fix12IiES2_S2_PK11Vector3_16f(
+unsigned int a, unsigned int b, int c, int d, int e, const Vector3_16f *f);
+void *_ZN8Particle6System12FromUniqueIDEj(unsigned int id);
+void func_ov006_020eef90(void);
+extern void* func_ov006_020c7300(void*);
+extern void* func_ov006_020c4060(void*);
+extern void func_ov004_020adb1c(int);
+extern void func_ov004_020b0a54(int);
+extern void func_ov006_020c42bc(void);
+extern void func_ov006_020c712c(void);
+extern int data_ov006_02140308;
+extern struct G2 data_ov006_0213cc74;
+extern struct G2 data_ov006_0213cc9c;
+extern void func_ov006_020c4148(void);
+extern void func_ov006_020c6f8c(int a);
+extern void func_ov006_020ef580(char *self);
+extern int data_ov006_02140304;
+extern int data_ov006_02140328;
+extern int data_ov006_02140428;
+extern struct G2 data_ov006_0213cc84;
+extern void func_ov006_020c7490(void);
+extern void func_ov006_020ef768(char *self);
+extern struct G2 data_ov006_0213cc94;
+extern int LoadFile(int handle);
+extern void DecompressLZ16(int src, int dst);
+extern void _ZN2GX11LoadOBJPlttEPKvjj(const void *p, u32 a, u32 b);
+extern s32 _ZN4cstd4fdivEii(s32 a, s32 b);
+extern void Camera_UpdateMatrices(void *cam);
+extern int func_ov006_020c4684(void *ptr, int n);
+extern int func_ov006_020c7574(void *base, int count);
+extern void *_ZN5Model8LoadFileER13SharedFilePtr(void *ref);
+extern int _ZN9ModelBase7SetFileEP8BMD_Fileii(void *self, void *bmdFile, int a, int b);
+extern u8 data_0209d45c;
+extern s16 data_02082614;
+typedef struct { int e[12]; } Mtx43T;
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020adb1c, from the legacy file for func_ov006_020ef5ac, NOT applied: extern void func_ov004_020adb1c(int self); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for _ZN12dScMgJump2_c13InitResourcesEv, NOT applied: extern s32 GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of Matrix4x3_FromTranslation, from the legacy file for _ZN12dScMgJump2_c13InitResourcesEv, NOT applied: extern void Matrix4x3_FromTranslation(void *m, int x, int y, int z); */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_021421b8, from the legacy file for _ZN12dScMgJump2_c13InitResourcesEv, NOT applied: extern int *data_ov006_021421b8; */
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 28 -- dScMgJump2_c_Spawn, 0x020efaf0, size 0x118 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void *_ZN7fBase_cnwEj(u32 sz);
+void _ZN11dScMgBase_cC2Ev(void *p);
+void func_020733a8(void *o, int a, int b, void *f1, void *f2);
+void _ZN8Particle10SysTrackerC1Ev(void *p);
+void _ZN5ModelC1Ev(void *p);
+extern char _ZTV14dScMgD3DBase_c[];
+void _ZN16dMgJump3DMario_cD1Ev();
+void _ZN16dMgJump3DMario_cC1Ev();
+void func_ov006_020c6f70();
+void func_ov006_020eed64();
+void func_ov006_020efc08();
+void func_ov006_020c6f3c();
+
+// @symbol dScMgJump2_c_Spawn
+/* recovered: vtable identified. Formerly mangled _ZN8PathLift17BaseInitResourcesEv --
+   a name matched across an OVERLAY BOUNDARY at a near-identical address to the real
+   PathLift (ov002, _ZN8PathLiftD1Ev and friends), neither this class's method nor a
+   BaseInitResources. This body is operator new(0x5a78), dScMgD3DBase_c's
+   constructor, dScMgD3DBase_c's vtable, mSysTracker, then dScMgJump2_c's own vtable
+   and every member -- a factory for dScMgJump2_c. Renamed per the fallback
+   convention (no known game object name for this minigame; compare
+   daObjEmmLog_c_Spawn, the tree's other `<ClassName>_Spawn` factory) rather than
+   the game-name-based MgBounceAndPounce_Spawn that dScMgJump_c's own factory
+   carries. See include/dScMgJump2_c.h. */
+void *dScMgJump2_c_Spawn()
+{
+    char *p = (char *)_ZN7fBase_cnwEj(0x5a78);
+    if (p) {
+        char *e;
+        char *end;
+        _ZN11dScMgBase_cC2Ev(p);
+        *(char **)p = _ZTV14dScMgD3DBase_c;
+        *(u16 *)(p + 0x4664) = 0;
+        e = p + 0x466c;
+        end = p;
+        end = end + 0x47e4;
+        do {
+            e += 0xbc;
+        } while (e != end);
+        _ZN8Particle10SysTrackerC1Ev(p + 0x47e4);
+        *(char **)p = (char *)&_ZTV12dScMgJump2_c[2];
+        func_020733a8(p + 0x500c, 3, 0xb8, (void *)_ZN16dMgJump3DMario_cC1Ev, (void *)_ZN16dMgJump3DMario_cD1Ev);
+        func_020733a8(p + 0x5234, 6, 0xf0, (void *)func_ov006_020c6f70, (void *)func_ov006_020c6f3c);
+        func_020733a8(p + 0x57d4, 0x10, 0x24, (void *)func_ov006_020efc08, (void *)func_ov006_020eed64);
+        _ZN5ModelC1Ev(p + 0x5a14);
+    }
+    return p;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 27 -- _ZN12dScMgJump2_c13OnYoshiTryEatEi, 0x020efaa8, size 0x48 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c13OnYoshiTryEatEi
+// recovered name: dScMgJump2_c_OnYoshiTryEat_020efaa8
+/* recovered: renamed to Class_Method, declarations from a shared header */
+/* recovered: renamed to Class_Method, vtable slot 18 -- an override of
+   dScMgBase_c::OnYoshiTryEat(int). The signature must repeat the base
+   declaration exactly, or mwcc appends a slot instead of overriding. */
+void dScMgJump2_c::OnYoshiTryEat(int /* arg */)
+{
+    char *c = (char *)this;
+
+  func_ov006_020c72b4();
+  func_ov006_020c719c(0, 0);
+  data_ov006_02140328 = 3;
+  func_ov006_020c44b4(0, 0);
+  func_ov006_020eeff0();
+  func_ov006_020ef7f8(c);
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 26 -- _ZN12dScMgJump2_c13OnTurnIntoEggEi, 0x020efa84, size 0x24 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c13OnTurnIntoEggEi
+// recovered name: dScMgJump2_c_OnTurnIntoEgg
+/* recovered: renamed to Class_Method, declarations from a shared header */
+/* recovered: renamed to Class_Method */
+/* dScMgJump2_c::OnTurnIntoEgg - recovered from vtable slot identity */
+int dScMgJump2_c::OnTurnIntoEgg(int /* mode */)
+{
+    func_ov006_020c8a9c(0, 0);
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 25 -- _ZN12dScMgJump2_c13InitResourcesEv, 0x020ef834, size 0x250 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c13InitResourcesEv
+/* dScMgJump2_c::InitResources -- vtable slot 0.
+ *
+ * Attributed by tools/rtti_vtables.py --own dScMgJump2_c, this class's own slot 0
+ * (fBase_c::InitResources). The old file's `recovered name:
+ * dScMgJump2_c_InitResources` agreed.
+ *
+ * The final `((Base *)base)->m48(-1)` is a self-dispatch through this class's own
+ * vtable slot 18 -- left as a raw vtable-shim call, same shape the pre-migration
+ * file used, just through `this` instead of a `char *base` parameter. */
+s32 dScMgJump2_c::InitResources()
+{
+    char *base = (char *)this;
+    int a, b;
+    s32 fov;
+    void *bmdFile;
+
+    a = LoadFile(data_ov006_0213cce0[GetGameLanguage()]);
+    b = LoadFile(0xd1);
+    DecompressLZ16(a, 0x6400000);
+    _ZN2GX11LoadOBJPlttEPKvjj((const void *)b, 0, 0x80);
+    Deallocate((void *)a);
+    Deallocate((void *)b);
+
+    data_0209d45c = 0x11;
+    _ZN3G3X6SetFogEbiii(0, 0, 2, 0x1000);
+    InitialiseVramGlobals();
+    *(u16 *)0x4000008 = (*(u16 *)0x4000008 & ~3) | 1;
+    fov = _ZN4cstd4fdivEii(0xc0000, (s32)data_02082614);
+
+    this->mCamera0EyeX = 0;
+    this->mCamera0EyeY = -0x64000;
+    this->mCamera0EyeZ = 0;
+    this->mCamera0TargetX = 0;
+    this->mCamera0TargetY = 0;
+    this->mCamera0TargetZ = fov;
+    this->mCamera0Angle = 0x800;
+    Camera_UpdateMatrices(base + 0x466c);
+
+    this->mCamera1EyeX = 0;
+    this->mCamera1EyeY = 0x82000;
+    this->mCamera1EyeZ = 0;
+    this->mCamera1TargetX = 0;
+    this->mCamera1TargetY = 0;
+    this->mCamera1TargetZ = fov;
+    this->mCamera1Angle = 0x800;
+    Camera_UpdateMatrices(base + 0x4728);
+
+    if (func_ov006_020c4684(base + 0x5234, 6) == 0)
+        return 0;
+
+    func_ov006_020ef0d4((int)(base + 0x57d4), 0x10);
+
+    data_ov006_02140314 = 1;
+    if (func_ov006_020c7574(base + 0x500c, 3) == 0)
+        return 0;
+
+    data_ov006_021421b8 = &data_ov006_021421c4;
+    bmdFile = _ZN5Model8LoadFileER13SharedFilePtr(&data_ov006_021421c4);
+
+    if (_ZN9ModelBase7SetFileEP8BMD_Fileii(base + 0x5a14, bmdFile, 1, -1) == 0) {
+        Matrix4x3_FromTranslation(&data_020a0e68, 0, 0, 0);
+        *(Mtx43T *)(base + 0x5a30) = *(Mtx43T *)&data_020a0e68;
+        return 0;
+    }
+
+    func_ov004_020b04d0(0x10);
+    ((Base *)base)->m48(-1);
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 24 -- func_ov006_020ef7f8, 0x020ef7f8, size 0x3c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef7f8(char* c){
+  *(short*)(c+0x5a74) = 0x78;
+  *(int*)(c+0x5a64) = 0;
+  *(int*)(c+0x5a68) = 0x4000;
+  *(struct G2*)(c+0x5004) = data_ov006_0213cc94;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 23 -- func_ov006_020ef794, 0x020ef794, size 0x64 */
+/* -------------------------------------------------------------------------- */
+/* func_ov006_020ef794 at 0x020ef794 (ov006), size 0x64
+ * Matched byte-for-byte with mwccarm 1.2/sp2p3.
+ * flags: -O4,p -enum int -lang c99 -char signed -interworking -proc arm946e -gccext,on -msgstyle gcc
+ */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef794(char *self)
+{
+    s16 *p = (s16 *)(((int)self + 0x5a74));
+    *p = *p - 1;
+    if (*(s16 *)(self + 0x5a74) == 0) {
+        if (*(u8 *)(self + 0xc4) == 0) {
+            *(u8 *)(self + 0xc3) = 1;
+            *(u8 *)(self + 0xc4) = 1;
+            *(s16 *)(self + 0xc0) = 0;
+        }
+        func_ov006_020c7490();
+        func_ov006_020ef768(self);
+    }
+    func_ov006_020c42bc();
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 22 -- func_ov006_020ef768, 0x020ef768, size 0x2c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef768(char *self)
+{
+    *(short *)(self + 0x5a74) = 0xf0;
+    *(struct G2 *)(self + 0x5004) = data_ov006_0213cc84;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 21 -- func_ov006_020ef5ac, 0x020ef5ac, size 0x1bc */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef5ac(char *self0)
+{
+    s16 flag;
+    char *self = self0;
+    int saved = data_ov006_02140308;
+    int *vp = (int *)((int)self + 0x5a64);
+    int c;
+
+    *vp += *(int *)(self + 0x5a68) >> 12;
+    if (*(int *)(self + 0x5a64) > 0x1000)
+        *vp -= 0x1000;
+
+    flag = 0;
+    c = data_ov006_02140304;
+    switch (c) {
+    case 0:
+        flag = 1;
+        break;
+    case 1:
+        {
+            int rc = _Z15ApproachLinear2Rsss((s16 *)(self + 0x5a74), flag, 2);
+            if (rc)
+                flag = 1;
+        }
+        break;
+    default:
+        {
+            int rc = _Z15ApproachLinear2Rsss((s16 *)(self + 0x5a74), flag, 1);
+            if (rc)
+                flag = 1;
+        }
+        break;
+    }
+
+    if (flag != 0 && data_ov006_02140304 < data_ov006_02140328) {
+        int t;
+        s16 v;
+        func_ov006_020c4148();
+        *(s16 *)(self + 0x5a74) = 0xf0 - data_ov006_02140308 * 6;
+        v = *(s16 *)(self + 0x5a74);
+        if (v < 0x3c)
+            v = 0x3c;
+        *(s16 *)(self + 0x5a74) = v;
+        t = (data_ov006_02140308 << 4) + 0x1600;
+        if (t > 0x2200)
+            t = 0x2200;
+        func_ov006_020c6f8c(t);
+    }
+
+    func_ov006_020c42bc();
+    func_ov006_020c712c();
+
+    if (saved != data_ov006_02140308) {
+        switch (data_ov006_02140308) {
+        case 3:
+            data_ov006_02140328 = 4;
+            *(char *)(self + 0xc3) = 0;
+            break;
+        case 0xa:
+            data_ov006_02140328 = 5;
+            break;
+        case 0x14:
+            data_ov006_02140328 = 6;
+            break;
+        }
+        func_ov004_020adb1c(data_ov006_02140308);
+    }
+
+    if (data_ov006_02140428 != 0)
+        return;
+    func_ov006_020ef580(self);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 20 -- func_ov006_020ef580, 0x020ef580, size 0x2c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef580(char *self)
+{
+    *(short *)(self + 0x5a74) = 0x20;
+    *(struct G2 *)(self + 0x5004) = data_ov006_0213cc9c;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 19 -- func_ov006_020ef4ec, 0x020ef4ec, size 0x94 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020c42bc(void);
+void func_ov006_020c712c(void);
+void func_ov006_020c7388(void);
+void func_ov006_020c40e8(void);
+extern struct G2 data_ov006_0213cc8c;
+void func_ov006_020ef4ec(char *c)
+{
+  int *p = (int *) ((unsigned long long) (c + 0x5a64));
+  char *b = c + 0x5000;
+  *p = (*p) + ((*((int *) (b + 0xa68))) >> 12);
+  if ((*((int *) ((c + 0x5000) + 0xa64))) > 0x1000)
+  {
+    *p = (*p) - 0x1000;
+  }
+  func_ov006_020c42bc();
+  func_ov006_020c712c();
+  if (_Z15ApproachLinear2Rsss((short *) (c + 0x5a74), 0, 1) == 0)
+  {
+    return;
+  }
+  func_ov006_020c7388();
+  func_ov006_020c40e8();
+  *((struct G2 *) (c + 0x5004)) = data_ov006_0213cc8c;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 18 -- func_ov006_020ef480, 0x020ef480, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef480(char *c){
+    void *r = func_ov006_020c7300(c);
+    if(r){
+        r = func_ov006_020c4060(r);
+        if(r){
+            func_ov004_020adb1c(data_ov006_02140308);
+            func_ov004_020b0a54(0x12);
+            *(unsigned char*)(c+0xc3)=0;
+            *(struct G2*)(c+0x5000+4) = data_ov006_0213cc74;
+        }
+    }
+    func_ov006_020c42bc();
+    func_ov006_020c712c();
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 17 -- func_ov006_020ef47c, 0x020ef47c, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef47c(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 16 -- _ZN12dScMgJump2_c8BehaviorEv, 0x020ef3e0, size 0x9c */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c8BehaviorEv
+/* dScMgJump2_c::Behavior -- vtable slot 6, ov006 0x020ef3e0.
+ *
+ * Attributed by the vtable: dScMgJump2_c's own table is ov006 0x0213ccfc and
+ * its slot 6 relocates here. The `recovered name:` comment on the
+ * pre-migration file agreed, but the vtable is what settles it. The signature
+ * is include/dScMgBase_c.h's own slot 6, `virtual s32 Behavior()`.
+ *
+ * unk_5a6c IS A PARTICLE SYSTEM UNIQUE ID, passed in and back out of
+ * Particle::System::NewUnkCallback818 each frame and then used to look the
+ * system up again. The two writes at +0x50 and +0x58 of whatever FromUniqueID
+ * returns are left as raw offsets: no header in this tree types that object.
+ *
+ * The two Particle::System declarations keep the local `unsigned int` /
+ * `int` spelling copied from the pre-migration file rather than the true
+ * Fix12<int> parameters their ROM names carry -- mwccarm passes by-value class
+ * parameters differently at the call site, so declaring the real types breaks
+ * the byte match. notes/mwccarm-codegen.md 6az records it; the sibling
+ * src/_ZN18dScMgTrampoline2_c8BehaviorEv.cpp says the same at its own site.
+ *
+ * THE STATE CALLBACK IS A POINTER-TO-MEMBER STORED IN THE OBJECT, in the
+ * eight bytes at 0x5004 that include/dScMgJump2_c.h calls pad_5004 -- eight
+ * being mwccarm's member-pointer size. The pre-migration file cast through a
+ * local incomplete `struct C` for the receiver; naming the real class here is
+ * byte-identical, which is what makes the field's extent evidence rather than
+ * a guess. */
+s32 dScMgJump2_c::Behavior()
+{
+    unk_5a6c = _ZN8Particle6System17NewUnkCallback818Ejj5Fix12IiES2_S2_PK11Vector3_16f(
+        unk_5a6c, 0xf0, 0x400000, 0x800000, -0x480000, 0);
+    void *o = _ZN8Particle6System12FromUniqueIDEj(unk_5a6c);
+    if (o != 0) {
+        *(int *)((char *)o + 0x50) = 0x4000;
+        *(unsigned char *)((char *)o + 0x58) = 0x2c;
+    }
+    func_ov006_020eef90();
+    {
+        dScMgJump2_cState *pp = (dScMgJump2_cState *)pad_5004;
+        (this->**pp)();
+    }
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 15 -- func_ov006_020ef2b8, 0x020ef2b8, size 0x128 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_0203cd80(int* m, int x);
+
+void func_ov006_020ef2b8(void) {
+    int v[3];
+    int w[3];
+    *(volatile int*)0x40004cc = 0x7fff;
+    *(volatile int*)0x40004cc = 0x40007fff;
+    v[0] = 0;
+    v[1] = 0;
+    v[2] = 0xfffff008;
+    func_0203cd80(v, 0x2000);
+    func_0203ccd4((struct Vector3 *)v, 0);
+    *(volatile int*)0x40004c8 =
+        (((short)v[0] >> 3) & 0x3ff) |
+        ((((short)v[1] >> 3) & 0x3ff) << 10) |
+        ((((short)v[2] >> 3) & 0x3ff) << 20);
+    w[2] = 0xfffff008;
+    w[0] = 0;
+    w[1] = 0;
+    *(volatile int*)0x4000440 = 2;
+    *(volatile int*)0x4000454 = 0;
+    func_0203cd80(w, 0x2000);
+    func_0203ccd4((struct Vector3 *)w, 0);
+    *(volatile int*)0x40004c8 =
+        ((((short)w[0] >> 3) & 0x3ff) |
+        ((((short)w[1] >> 3) & 0x3ff) << 10) |
+        ((((short)w[2] >> 3) & 0x3ff) << 20)) | 0x40000000;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 14 -- _ZN12dScMgJump2_c6RenderEv, 0x020ef148, size 0x170 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c6RenderEv
+/* dScMgJump2_c::Render -- vtable slot 9, ov006 0x020ef148.
+ *
+ * Attributed by the vtable: dScMgJump2_c's own table is ov006 0x0213ccfc and
+ * its slot 9 relocates here. The signature is include/dScMgBase_c.h's own slot
+ * 9, `virtual s32 Render()`.
+ *
+ * THE TWO STRUCT COPIES ARE THE ONE THING IN THIS FILE THAT IS NOT FREE, and
+ * S3 HAD TO CHANGE SHAPE TO SURVIVE THE MOVE TO //cpp. A whole-struct
+ * assignment is a memcpy in C and is scalarised in C++ (notes/
+ * mwccarm-codegen.md), and the pre-migration file's `struct S3 { int a, b, c; }`
+ * duly came out three ldr/str pairs where the ROM has one ldm/stm -- 0x17c
+ * against 0x170, twelve bytes, exactly the defect that note names. Declaring
+ * the same three words as an ARRAY member restores the block copy. It is a
+ * size threshold and not a language-mode absolute: M48, twelve words, block-
+ * copies in C++ either way, which is why the 0x5a30 matrix store needed
+ * nothing. Both stay anonymous word arrays rather than named matrix or vector
+ * types -- spelling either as the real type is a separate change that has to
+ * be re-measured against this same gate.
+ *
+ * M48 is the 12-word Matrix4x3 the ROM writes into mModel at +0x1c; S3 is the
+ * three-word argument the model's own slot-5 virtual takes.
+ *
+ * mModel is raw bytes here for the reason include/dScMgJump2_c.h's banner
+ * gives -- the ROM's destructor destroys it before the three arrays, which a
+ * typed member could never reproduce -- so its insides are reached by offset:
+ * +0x00 is the vptr the last call dispatches through, +0x08 the object
+ * func_02045e44 steps, +0x1c the matrix. */
+s32 dScMgJump2_c::Render()
+{
+    char *self = (char *)this;
+    struct S3 local;
+    void *p;
+    int count;
+    int a1v;
+    int i;
+
+    func_ov004_020b1a5c(data_ov006_02140308, 4);
+    func_ov006_020eef58();
+    func_ov006_020ef2b8();
+    Matrix4x3_FromTranslation(&data_020a0e68, 0, 0xffead000, -0xa0000);
+    *(struct M48 *)(self + 0x5a30) = *(struct M48 *)&data_020a0e68;
+
+    local = *(struct S3 *)&data_ov006_0213ccb0;
+
+    {
+        int *obj = (int *)(self + 0x5a1c);
+        p = (void *)obj[1];
+        func_02045e44((void *)obj, 1, 0);
+    }
+    *(int *)((char *)p + 0x18) = unk_5a64;
+    {
+        void **vobj = (void **)(self + 0x5a14);
+        (*(void (**)(void *, void *))((char *)*vobj + 0x14))((void *)vobj, &local);
+    }
+
+    func_ov006_020c70d0();
+    func_ov006_020c425c();
+
+    if (*(u16 *)(self + 0x4664) == 1) {
+        a1v = 0x6e;
+        count = data_ov006_02140428;
+        for (i = 0; i < 3; i++) {
+            if (i >= count) {
+                int idx = GetGameLanguage();
+                DrawOamSprite((void *)data_ov006_0213cccc[idx][1], (void *)a1v, 0xc, (void *)0);
+            } else {
+                RenderOamMainScreen(data_ov006_02134cf8, a1v, 0xc, -1, -1);
+            }
+            a1v += 0x12;
+        }
+    }
+
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 13 -- _ZN12dScMgJump2_c16CleanupResourcesEv, 0x020ef110, size 0x38 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_c16CleanupResourcesEv
+/* dScMgJump2_c::CleanupResources -- vtable slot 3, ov006 0x020ef110.
+ *
+ * Attributed by the vtable: dScMgJump2_c's own table is ov006 0x0213ccfc (the
+ * address include/dScMgJump2_c.h's banner records) and its slot 3 relocates
+ * here, so this is this class's own override and not dScMgD3DBase_c's. The
+ * signature is include/fBase_c.h's own slot 3, `virtual s32
+ * CleanupResources()`.
+ *
+ * data_ov006_021421b8 IS A POINTER TO the handle, not the handle: the ROM
+ * loads the word and passes its VALUE, then stores 0 back. That is why it is
+ * spelled `void *` here and cast at the call, unlike dScMg3DEsp_c's own
+ * CleanupResources where the two globals ARE the handles and their addresses
+ * are the arguments.
+ *
+ * Release() comes from include/SharedFilePtr.h rather than from a bare
+ * `extern void _ZN13SharedFilePtr7ReleaseEv(void *)`: in a //cpp translation
+ * unit that spelling mangles a SECOND time and emits a reference nothing
+ * defines, which is the defect that header's own banner records. */
+s32 dScMgJump2_c::CleanupResources()
+{
+    ((SharedFilePtr *)data_ov006_021421b8)->Release();
+    data_ov006_021421b8 = 0;
+    func_ov004_020ad90c();
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 12 -- func_ov006_020ef0d4, 0x020ef0d4, size 0x3c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef0d4(int a, int b){
+  data_ov006_021421b0 = (struct Elem *)a;
+  data_ov006_021421bc = b;
+  data_ov006_021421b4 = 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 11 -- func_ov006_020ef05c, 0x020ef05c, size 0x78 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020ef05c(int p0, int p1, int p2)
+{
+    int i;
+    for (i = data_ov006_021421bc - 1; i >= 0; i--) {
+        if (data_ov006_021421b0[i].f20 == 0) {
+            func_ov006_020eee3c((int *)&data_ov006_021421b0[i], p0, p1, p2);
+            return;
+        }
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 10 -- func_ov006_020eeff0, 0x020eeff0, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eeff0(void){
+    int i;
+    for(i=0;i<data_ov006_021421bc;i++){
+        *(short*)((char *)data_ov006_021421b0 + i*0x24 + 0x20)=0;
+    }
+    data_ov006_021421c0=0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 9 -- func_ov006_020eef90, 0x020eef90, size 0x60 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eef90(void)
+{
+    int i = 0;
+    if (i < data_ov006_021421bc) {
+        int off = 0;
+        do {
+            func_ov006_020eedc8((char *)data_ov006_021421b0 + off);
+            i++;
+            off += 0x24;
+        } while (i < data_ov006_021421bc);
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 8 -- func_ov006_020eef58, 0x020eef58, size 0x38 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eef58(void) {
+    void *node = data_ov006_021421c0;
+    if (node == 0) return;
+    do {
+        func_ov006_020eed68((char *)node);
+        node = *(void **)node;
+    } while (node != 0);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 7 -- func_ov006_020eef40, 0x020eef40, size 0x18 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eef40(void *node){
+  *(void**)node = data_ov006_021421c0;
+  data_ov006_021421c0 = (struct Node *)node;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 6 -- func_ov006_020eeef4, 0x020eeef4, size 0x4c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eeef4(struct Node* node) {
+    struct Node* prev = data_ov006_021421c0;
+    if (prev == node) {
+        data_ov006_021421c0 = node->next;
+        return;
+    }
+    struct Node* cur = prev->next;
+    while (cur) {
+        if (cur == node) {
+            prev->next = node->next;
+            return;
+        }
+        prev = cur;
+        cur = cur->next;
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 5 -- func_ov006_020eee3c, 0x020eee3c, size 0xb8 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov004_020b04c0(void);
+void func_ov006_020eef40(void *node);
+
+void func_ov006_020eee3c(int *thiz, int a1, int a2, int a3)
+{
+    thiz[4/4] = a1;
+    thiz[8/4] = a2;
+    thiz[0x14/4] = 0;
+    thiz[0x18/4] = 0x4000;
+    thiz[0xc/4] = a1;
+    thiz[0x10/4] = a2 - 0x40000;
+    if (thiz[0x10/4] < 0x8000) {
+        int t = func_ov004_020b04c0() << 12;
+        if (thiz[0x10/4] > -t) {
+            thiz[0x10/4] = 0x8000;
+            goto strh;
+        }
+    }
+    if (thiz[0x10/4] >= -0xb8000 - (func_ov004_020b04c0() << 12))
+        goto strh;
+    thiz[0x10/4] = -0xb8000 - (func_ov004_020b04c0() << 12);
+strh:
+    *(short*)((char*)thiz + 0x1c) = 0x80;
+    *(short*)((char*)thiz + 0x20) = 1;
+    *(short*)((char*)thiz + 0x1e) = (short)a3;
+    func_ov006_020eef40(thiz);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 4 -- func_ov006_020eedc8, 0x020eedc8, size 0x74 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eedc8(char* c) {
+    if (*(short*)(c + 0x20) == 0) return;
+    if (_Z15ApproachLinear2Rsss((short*)(c + 0x1c), 0, 1) != 0) {
+        func_ov006_020eeef4((struct Node *)c);
+        *(short*)(c + 0x20) = 0;
+    }
+    _Z14ApproachLinearRiii((int*)(c + 4), *(int*)(c + 0xc), *(int*)(c + 0x14));
+    _Z14ApproachLinearRiii((int*)(c + 8), *(int*)(c + 0x10), *(int*)(c + 0x18));
+    func_0203d630((int*)(c + 0x14), 0xf00);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 3 -- func_ov006_020eed68, 0x020eed68, size 0x60 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eed68(char *c)
+{
+    if (*(short *)(c + 0x20) == 0) return;
+    func_ov004_020b2444(*(int *)(c + 4) >> 12, *(int *)(c + 8) >> 12,
+                        *(short *)(c + 0x1e), -1, -1, 0, data_ov006_021421b4);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 2 -- func_ov006_020eed64, 0x020eed64, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020eed64(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 1 -- _ZN12dScMgJump2_cD0Ev, 0x020eec9c, size 0xc8 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_cD0Ev
+/* NO BODY HERE, DELIBERATELY. The deleting destructor D0 is not a second
+   function to write: mwccarm generates it from the one out-of-line
+   ~dScMgJump2_c() below, which the legacy per-function split had to duplicate
+   because each file could hold only one symbol. Writing it twice in one
+   translation unit is a redefinition. */
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 0 -- _ZN12dScMgJump2_cD1Ev, 0x020eebe8, size 0xb4 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN12dScMgJump2_cD1Ev
+/* recovered: real C++ destructor. The explicit Model and untyped-array calls
+   reproduce the first three steps of the ROM body. The compiler then destroys
+   mPlayers as the final derived member. Everything after -- this class's vtable
+   store, dScMgD3DBase_c's vptr store, mSysTracker destruction and the chain
+   to ~dScMgBase_c() -- is the compiler's own inlining of
+   dScMgD3DBase_c's inline destructor (see include/dScMgD3DBase_c.h). The Model
+   call remains explicit because the ROM destroys it before both arrays. */
+/* NO BODY HERE. The destructor is defined in the class body in
+   include/dScMgJump2_c.h -- see the long note there for why the out-of-line
+   spelling cannot be linked as one translation unit. mwccarm emits D1 and D0
+   from that one definition, in this order, and emits no D2. */
