--- src/func_ov006_020eef40.c
@@ -1,5 +1,0 @@
-extern void *data_ov006_021421c0;
-void func_ov006_020eef40(void *node){
-  *(void**)node = data_ov006_021421c0;
-  data_ov006_021421c0 = node;
-}
