--- src/func_ov006_020ef0d4.c
@@ -1,8 +1,0 @@
-extern int data_ov006_021421b0;
-extern int data_ov006_021421bc;
-extern int data_ov006_021421b4;
-void func_ov006_020ef0d4(int a, int b){
-  data_ov006_021421b0 = a;
-  data_ov006_021421bc = b;
-  data_ov006_021421b4 = 0;
-}
