--- src/func_ov006_020ef47c.c
@@ -1,3 +1,0 @@
-void func_ov006_020ef47c(void)
-{
-}
